forgot to mention in previous commit
Ragdolling was added on death of unit this is done by adding a ragdoll componenet then on death using a recursive call to loop through all the ragdoll bones and match their positions and rotations to the unit that died on death to make a more natural death animation instead of spawning in a tpose the ragdoll spawns with same pos. Also an explosion force are added (hit and shoot effects also have vfxs hadded, some from assets and some are custom).